+++ CHANGELOG.rst
@@ -1,6 +1,14 @@
-**3.0.11 - 10/08/24**
+**3.0.13 - 10/15/24**
 
   - Fix mypy errors: vivarium/framework/lookup/interpolation.py
+
+**3.0.12 - 10/14/24**
+
+  - Bugfix for mypy errors: vivarium/framework/values.py
+
+**3.0.11 - 10/08/24**
+
+  - Fix mypy errors: vivarium/framework/values.py
 
 **3.0.10 - 10/07/24**